p2-metadata-ingestion/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3

# Database
sqlalchemy==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Task queue
celery==5.5.1
redis==5.3.0

# Storage
boto3==1.38.5
botocore==1.38.5

# File processing
python-magic==0.4.27
python-multipart==0.0.20

# Observability
prometheus-client==0.21.1
